backend/requirements.txt
# =============================================================================
# Synergy Sales Genius - Python Backend Dependencies
# Pin versions for reproducible enterprise deployments.
# Install: pip install -r requirements.txt
# =============================================================================

# --- Web Framework ---
fastapi==0.115.5
uvicorn[standard]==0.32.1

# --- Data Validation ---
pydantic==2.10.3
pydantic-settings==2.6.1

# --- Azure OpenAI SDK (new unified client) ---
openai==1.57.0

# --- Azure Cosmos DB SDK ---
azure-cosmos==4.9.0

# --- Environment & Utilities ---
python-dotenv==1.0.1
numpy==2.2.0        # Used for cosine similarity in duplicate detection

# --- CORS & HTTP ---
httpx==0.28.0       # Async HTTP client used internally by openai SDK
python-multipart==0.0.22  # Required by FastAPI for file upload (UploadFile / Form) endpoints

# --- Authentication (Phase 2) ---
python-jose[cryptography]==3.3.0   # JWT encode/decode (jose.jwt)
cryptography==43.0.3               # Required by python-jose[cryptography] for RSA/EC
passlib[bcrypt]==1.7.4             # bcrypt password hashing for Cosmos DB Users container
bcrypt==4.2.1                      # Underlying bcrypt implementation used by passlib

# --- Rate Limiting ---
slowapi==0.1.9                     # IP-based rate limiting middleware for FastAPI (wraps `limits`)
limits==3.14.0                     # Backend library used by slowapi for rate counters

# --- Azure Monitor / Application Insights ---
# Provides OpenTelemetry auto-instrumentation: request traces, dependency tracking,
# exception telemetry, log correlation, and live metrics streamed to App Insights.
# No-op when APPLICATIONINSIGHTS_CONNECTION_STRING env var is absent (local dev).
azure-monitor-opentelemetry==1.6.4
